feat: gate vault/asset growth by EIP-7825 gas feasibility domain

+++ contracts/test/LiquidityOrchestratorVaultHarness.sol
@@ -24,7 +24,7 @@ contract LiquidityOrchestratorVaultHarness is LiquidityOrchestrator {
         minibatchSize = size;
     }
 
-    function h_setCurrentMinibatchIndex(uint8 index) external {
+    function h_setCurrentMinibatchIndex(uint16 index) external {
         currentMinibatchIndex = index;
     }
